requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==4.2.19
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
greenlet==3.1.1
numpy==2.0.2
pandas==2.3.0
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
